- Extracted logic to separate component for reusability.
