+++ src/app/[version]/[alloy]/page.tsx
@@ -1,19 +1,14 @@
 "use client";
 
-import {useRouter, useParams} from "next/navigation";
+import {useParams} from "next/navigation";
 import {AlloyComponentDisplay} from "@/components/AlloyComponentDisplay";
-import {CaretCircleLeft} from "@phosphor-icons/react";
+import {HeadingWithBackButton} from "@/components/HeadingWithBackButton";
 
 
 export default function AlloyPage() {
-	const router = useRouter();
 	const params = useParams();
 	const version = params.version as string;
 	const alloy = params.alloy as string;
-
-	const handleBack = () => {
-		router.push(`/${version}/alloys`);
-	};
 
 	return (
 			<main
@@ -22,26 +17,11 @@ export default function AlloyPage() {
 					aria-label="Alloy Calculator"
 			>
 				<div className="max-w-6xl mx-auto">
-					<div className="flex items-center mb-20">
-						<button
-								onClick={handleBack}
-								className="mr-4"
-								aria-label="Return to alloy selection"
-						>
-							<CaretCircleLeft
-									size={40}
-									weight="bold"
-									className="text-primary text-teal-100 hover:text-teal-300 transition-colors duration-200"
-									aria-hidden="true"
-							/>
-						</button>
-
-						<h1 className="text-3xl font-bold text-primary flex-grow text-center">
-							{alloy.toUpperCase()}
-						</h1>
-
-						<div className="w-10" aria-hidden="true"/>
-					</div>
+					<HeadingWithBackButton
+							title={alloy.toUpperCase()}
+							ariaPreviousScreenName="alloy selection"
+							handleBackURI={`/${version}/alloys`}
+					/>
 
 					<AlloyComponentDisplay alloy={alloy}/>
 				</div>
